@@ -4,57 +4,57 @@ publish_to: 'none'
 version: 3.4.2+41
 
 environment:
-  sdk: ">=3.2.0 <4.0.0"
-  flutter: ">=3.16.0 <4.0.0"
+  sdk: ">=3.4.0 <4.0.0"
+  flutter: ">=3.22.0 <4.0.0"
 
 dependencies:
   auto_size_text: ^3.0.0
-  cached_network_image: ^3.3.0
+  cached_network_image: ^3.3.1
   collection: ^1.18.0
   crypto: ^3.0.3
   datetime_picker_formfield: ^2.0.1
-  fl_chart: ^0.64.0
+  fl_chart: ^0.68.0
   flutter:
     sdk: flutter
   flutter_acrcloud: ^0.2.1
-  flutter_cache_manager: ^3.3.1
-  flutter_local_notifications: ^16.1.0
+  flutter_cache_manager: ^3.3.2
+  flutter_local_notifications: ^17.1.2
   flutter_mpmediaplayer: ^0.3.0
   flutter_web_auth: ^0.5.0
   gallery_saver: ^2.3.2
   html: ^0.15.4
-  http: ^1.1.0
-  in_app_review: ^2.0.8
-  intl: ^0.18.1
-  json_annotation: ^4.8.1
-  modal_bottom_sheet: ^2.1.2
+  http: ^1.2.1
+  in_app_review: ^2.0.9
+  intl: ^0.19.0
+  json_annotation: ^4.9.0
+  modal_bottom_sheet: ^3.0.0
   octo_image: ^2.0.0
-  package_info_plus: ^4.2.0
-  path_provider: ^2.1.1
+  package_info_plus: ^8.0.0
+  path_provider: ^2.1.3
   pkce: ^1.1.0+1
   pool: ^1.5.1
-  quick_actions: ^1.0.6
+  quick_actions: ^1.0.7
   rxdart: ^0.27.7
-  share_plus: ^7.2.1
-  shared_preferences: ^2.2.2
+  share_plus: ^9.0.0
+  shared_preferences: ^2.2.3
   uni_links: ^0.5.1
-  sqflite: ^2.3.0
+  sqflite: ^2.3.3+1
   universal_html: ^2.2.4
   universal_io: ^2.2.2
-  url_launcher: ^6.2.1
+  url_launcher: ^6.2.6
   visibility_detector: ^0.4.0+2
   wakelock: ^0.6.2
   workmanager: ^0.5.2
 
 dev_dependencies:
-  build_runner: ^2.4.6
+  build_runner: ^2.4.9
   flutter_launcher_icons: ^0.13.1
-  flutter_lints: ^3.0.1
+  flutter_lints: ^4.0.0
   flutter_test:
     sdk: flutter
   integration_test:
     sdk: flutter
-  json_serializable: ^6.7.1
+  json_serializable: ^6.8.0
 
 dependency_overrides:
   flutter_launcher_icons:
@@ -63,11 +63,6 @@ dependency_overrides:
     git:
       url: https://github.com/Zazo032/gallery_saver.git
       ref: patch-1
-  modal_bottom_sheet:
-    git:
-      url: https://github.com/jamesblasco/modal_bottom_sheet.git
-      path: modal_bottom_sheet
-      ref: feat/sheet_improvements
   wakelock:
     git:
       url: https://github.com/formigas/wakelock.git
