@@ -70,7 +70,7 @@ class _ScoreboardState extends State<Scoreboard> {
       .map((key) => widget.statisticActions.containsKey(key)
           ? OutlinedButton(
               style: ButtonStyle(
-                  side: MaterialStateProperty.all(BorderSide(
+                  side: WidgetStateProperty.all(BorderSide(
                       color: Theme.of(context).colorScheme.primary))),
               onPressed: widget.statisticActions[key],
               child: _scoreTile(key),
